refactor(components): optimize code naming
